+++ src/environments/environment.ts
@@ -16,12 +16,4 @@ export const environment = {
 // import 'zone.js/dist/zone-error';  // Included with Angular CLI.
 // For Firebase JS SDK v7.20.0 and later, measurementId is optional
 
-export const firebaseConfig = {
-  apiKey: 'AIzaSyCd17MV0EujR5kNu25uz7F_RH1r8tdbjDs',
-  authDomain: 'linkpacket.firebaseapp.com',
-  projectId: 'linkpacket',
-  storageBucket: 'linkpacket.appspot.com',
-  messagingSenderId: '527170168300',
-  appId: '1:527170168300:web:94f41f89652e9999c90cc9',
-  measurementId: 'G-TTJDH8RS30'
-};
+
