+++ src/forFirebase/config.spec.ts
@@ -1,0 +1,7 @@
+import { Config } from './config';
+
+describe('Config', () => {
+  it('should create an instance', () => {
+    expect(new Config()).toBeTruthy();
+  });
+});
